blocks-tab renamed to block-types-tab in inserter menu
